Struct adsensehost4d1::AccountMethodsBuilder

pub struct AccountMethodsBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on account resources. It is not used directly, but through the AdSenseHost hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate adsensehost4d1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdSenseHost::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `adclients_get(...)`, `adclients_list(...)`, `adunits_delete(...)`, `adunits_get(...)`, `adunits_get_ad_code(...)`, `adunits_insert(...)`, `adunits_list(...)`, `adunits_patch(...)`, `adunits_update(...)`, `get(...)`, `list(...)` and `reports_generate(...)`
// to build up your call.
let rb = hub.accounts();

Methods

impl<'a, C, NC, A> AccountMethodsBuilder<'a, C, NC, A>

fn adunits_list(&self, account_id: &str, ad_client_id: &str) -> AccountAdunitListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

List all ad units in the specified publisher's AdSense account.

fn list(&self, filter_ad_client_id: &str) -> AccountListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

List hosted accounts associated with this AdSense account by ad client id.

fn adunits_get_ad_code(&self, account_id: &str, ad_client_id: &str, ad_unit_id: &str) -> AccountAdunitGetAdCodeMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get ad code for the specified ad unit, attaching the specified host custom channels.

fn reports_generate(&self, account_id: &str, start_date: &str, end_date: &str) -> AccountReportGenerateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Generate an AdSense report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

fn adunits_delete(&self, account_id: &str, ad_client_id: &str, ad_unit_id: &str) -> AccountAdunitDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Delete the specified ad unit from the specified publisher AdSense account.

fn adclients_list(&self, account_id: &str) -> AccountAdclientListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

List all hosted ad clients in the specified hosted account.

fn adunits_patch(&self, request: &AdUnit, account_id: &str, ad_client_id: &str, ad_unit_id: &str) -> AccountAdunitPatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Update the supplied ad unit in the specified publisher AdSense account. This method supports patch semantics.

fn adclients_get(&self, account_id: &str, ad_client_id: &str) -> AccountAdclientGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get information about one of the ad clients in the specified publisher's AdSense account.

fn get(&self, account_id: &str) -> AccountGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get information about the selected associated AdSense account.

fn adunits_insert(&self, request: &AdUnit, account_id: &str, ad_client_id: &str) -> AccountAdunitInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Insert the supplied ad unit into the specified publisher AdSense account.

fn adunits_get(&self, account_id: &str, ad_client_id: &str, ad_unit_id: &str) -> AccountAdunitGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get the specified host ad unit in this AdSense account.

fn adunits_update(&self, request: &AdUnit, account_id: &str, ad_client_id: &str) -> AccountAdunitUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Update the supplied ad unit in the specified publisher AdSense account.